@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IdentityModel.Tokens.Jwt;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Autofac;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
@@ -12,10 +7,14 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
-using RestaurantApp.WebApi.Entities;
+using Microsoft.OpenApi.Models;
+using RestaurantApp.DataServices;
+using RestaurantApp.WebApi.Infrastructure;
+using System;
+using System.IdentityModel.Tokens.Jwt;
+using System.Reflection;
+using System.Text;
 
 namespace RestaurantApp.WebApi
 {
@@ -31,12 +30,15 @@ namespace RestaurantApp.WebApi
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            services.AddSwaggerGen(c =>
+            {
+                c.SwaggerDoc("v1", new OpenApiInfo { Title = "RestaurantApp.WebAPI", Version = "v1" });
+            });
 
-            services.AddDbContext<ApplicationDbContext>(options => options.UseSqlServer(Configuration.GetConnectionString("RestaurantDatabase")));
+            services.AddDbContext<Entities>(options => options.UseSqlServer(Configuration.GetConnectionString("RestaurantDatabase")));
 
             services.AddIdentity<IdentityUser, IdentityRole>()
-                .AddEntityFrameworkStores<ApplicationDbContext>()
+                .AddEntityFrameworkStores<Entities>()
                 .AddDefaultTokenProviders();
 
             services.Configure<IdentityOptions>(options =>
@@ -55,15 +57,17 @@ namespace RestaurantApp.WebApi
                 options.SuppressModelStateInvalidFilter = true;
             });
 
-            services.AddCors(options =>
-            {
-                options.AddPolicy("CorsPolicy",
-                    builder => builder.AllowAnyOrigin()
-                      .AllowAnyMethod()
-                      .AllowAnyHeader()
-                      .AllowCredentials()
-                .Build());
-            });
+            services.AddControllers();
+
+            //services.AddCors(options =>
+            //{
+            //    options.AddPolicy("CorsPolicy",
+            //        builder => builder.AllowAnyOrigin()
+            //          .AllowAnyMethod()
+            //          .AllowAnyHeader()
+            //          .AllowCredentials()
+            //    .Build());
+            //});
 
             // ===== Add Jwt Authentication ========
             JwtSecurityTokenHandler.DefaultInboundClaimTypeMap.Clear(); // => remove default claims
@@ -88,31 +92,41 @@ namespace RestaurantApp.WebApi
                         RequireExpirationTime = true
                     };
                 });
+
+            var builder = new ContainerBuilder();
+            builder.RegisterAssemblyModules(Assembly.GetExecutingAssembly());
+            builder.RegisterModule(new WebApiModule());
+            var container = builder.Build();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app,
-            IHostingEnvironment env,
-            ApplicationDbContext dbContext,
-            UserManager<IdentityUser> userManager,
-            RoleManager<IdentityRole> roleManager)
+            IHostingEnvironment env)
+        //Entities dbContext,
+        //UserManager<IdentityUser> userManager,
+        //RoleManager<IdentityRole> roleManager)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
+                app.UseSwagger();
+                app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "RestaurantApp.WebAPI v1"));
             }
 
-            dbContext.Database.EnsureCreated();
+            //dbContext.Database.EnsureCreated();
 
-            IdentityInitializer.SeedData(userManager, roleManager);
+            //IdentityInitializer.SeedData(userManager, roleManager);
 
             app.UseStaticFiles();
 
             app.UseAuthentication();
 
             app.UseCors("CorsPolicy");
-
-            app.UseMvc();                
+            app.UseRouting();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
         }
     }
 }
